[Refactor] Separate the color utility
-[Refactor] Refactor color slider get style function
